Name GpEnemy::field_42 and correct a func_8009EA50 prototype
Two shared-code items surfaced by the actor typing passes.

GpEnemy had a 9-byte pad at 0x42, but func_actor_405800_8013340C reads a
halfword there and compares field_40 against quarters and eighths of it, so it
is the max-HP ceiling field_40 is measured against. Named as u16, with the pad
shortened to 7 bytes so the struct size is unchanged. This was the one function
that overlay had to skip.

src/actors/lib/actors_shared_80137a20.c declared func_8009EA50 taking s16 while
src/gameplay/gameplay.c:3757 defines it taking s32 - a real disagreement that
would sign-extend differently through the shared unit's call. Corrected to the
definition's s32.

Verified with the bare unscoped build.

+++ src/actors/lib/actors_shared_80137a20.c
@@ -4,7 +4,7 @@
 #include "actors/actors_shared_80137b78.h"
 
 void Gp_UpdateCoord(GsCOORDINATE2* arg0);
-void func_8009EA50(s16 arg0);
+void func_8009EA50(s32 arg0); /* defined (s32) in src/gameplay/gameplay.c */
 
 void ActorsShared80137a20_Fn31F54(ActorShared80137a20* arg0, ActorShared80137a20Obj2C* arg1, s32 arg2);
 void ActorsShared80137a20_Fn34968(ActorShared80137a20* arg0);
